packages/core/pyproject.toml
[project]
name = "db-mcp"
version = "0.8.12"
description = "MCP server for database semantics and query intelligence"
requires-python = ">=3.13"
dependencies = [
    "pydantic-ai>=0.1.0",
    "pydantic-graph>=0.1.0",
    "pydantic-settings>=2.0.0",
    "sqlalchemy>=2.0.0",
    "sqlglot>=26.0.0",
    "trino>=0.330.0",
    "clickhouse-sqlalchemy>=0.3.0",
    "psycopg2-binary>=2.9.0",
    "pymysql>=1.1.0",
    "pymssql>=2.2.0",
    "pyyaml>=6.0.0",
    "db-mcp-models",
    "db-mcp-knowledge",
    "db-mcp-data",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "bicp-agent",
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "dulwich>=0.22.0",
    "duckdb>=1.0.0",
]

[project.scripts]
db-mcp-benchmark = "db_mcp.benchmark:main"

[tool.uv.sources]
db-mcp-models = { workspace = true }
db-mcp-knowledge = { workspace = true }
db-mcp-data = { workspace = true }
bicp-agent = { git = "https://github.com/apelogic-ai/bicp.git", subdirectory = "packages/agent-python", rev = "743ecc5" }

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
    "pyinstaller>=6.0.0",
    "db-mcp-cli",
    "db-mcp-server",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pyinstaller>=6.18.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.13",
]

[tool.ruff]
line-length = 99
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "e2e: end-to-end integration tests using real git repos (slower)",
]
